+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `Builder::chunk()` now respects pre-defined `limit` and `offset` constraints
+
 ## [1.1.0] - 2025-11-21
 
 ### Added